requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
celery[redis]==5.4.0
redis==5.0.8
anthropic>=0.26.0
httpx==0.27.2
pydantic-settings==2.4.0
PyYAML==6.0.2
structlog==24.4.0
semgrep==1.86.0
pytest==8.3.3
pytest-asyncio==0.24.0
respx==0.21.1
